+++ package.json
@@ -11,7 +11,7 @@
     "access": "public"
   },
   "resolutions": {
-    "graphql": "14.5.8"
+    "graphql": "15.5.0"
   },
   "devDependencies": {
     "@types/convict": "^4.2.0",
@@ -37,10 +37,17 @@
   "dependencies": {
     "@koa/cors": "^2.2.2",
     "@koa/router": "^9.4.0",
+    "@opentelemetry/exporter-collector-grpc": "^0.16.0",
+    "@opentelemetry/exporter-prometheus": "^0.16.0",
+    "@opentelemetry/instrumentation-graphql": "^0.13.0",
+    "@opentelemetry/instrumentation-http": "^0.16.0",
+    "@opentelemetry/koa-instrumentation": "^0.13.0",
+    "@opentelemetry/sdk-node": "^0.16.0",
     "@types/jsonwebtoken": "^8.5.0",
     "@types/koa__router": "^8.0.3",
     "convict": "^4.2.0",
     "dataloader": "^1.4.0",
+    "graphql": "^15.5.0",
     "graphql-tools": "^4.0.5",
     "jsonwebtoken": "^8.5.1",
     "jwks-rsa": "^1.9.0",
