+++ src/client-graphql/modules/default.ts
@@ -13,7 +13,7 @@ export async function createModuleType(
   usedTypeNames: Set<string>,
   tableByName: TableByName
 ): Promise<GraphQLObjectType> {
-  const fields: GraphQLFieldConfigMap<unknown, unknown, unknown> = {};
+  const fields: GraphQLFieldConfigMap<unknown, unknown> = {};
   for (const [n, v] of Object.entries(tableByName)) {
     if (n !== "") {
       const tableFieldName = toSafeName(n);
