+++ src/client-graphql/modules/properties.ts
@@ -23,7 +23,7 @@ export async function createModuleType(
   usedTypeNames: Set<string>,
   tableByName: TableByName
 ): Promise<GraphQLObjectType> {
-  const fields: GraphQLFieldConfigMap<unknown, unknown, unknown> = {};
+  const fields: GraphQLFieldConfigMap<unknown, unknown> = {};
   const propertyTable = tableByName["property"];
 
   const propertyValueTranslationRowType = new GraphQLObjectType({
