+++ src/client-graphql/modules/sound.ts
@@ -14,7 +14,7 @@ export async function createModuleType(
   usedTypeNames: Set<string>,
   tableByName: TableByName
 ): Promise<GraphQLObjectType> {
-  const fields: GraphQLFieldConfigMap<unknown, unknown, unknown> = {};
+  const fields: GraphQLFieldConfigMap<unknown, unknown> = {};
   const soundVariantTable = tableByName["sound_variant"];
 
   const damperRowType = new GraphQLObjectType({
